+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adding the `phonetics/french/soundex` namespace.
 * Adding the `phonetics/french/soundex2` namespace.
 * Adding the `phonetics/french/phonex` namespace.
+* Adding the `phonetics/french/sonnex` namespace.
 * Adding the `stemmers/french/porter` namespace.
 * Adding the `tokenizers/sentences/punkt` namespace.
 * Moving the Cologne phonetic algorithm to the `phonetics/german/cologne` namespace.